Make changes to HTML to apply CSS font and color
Add header-logo class and hero class to separate from the rest
of the header and apply different color and font properties. Add
thor class to apply different properties than the rest of the
quote class. Add <h3> to "call to action" section to apply
different properites from the rest of the call-section. Fix
typos.