Add classes to HTML to apply text and color CSS
Add a header to "some information" in order to apply different
size and color properties than the subtext below the images.
Add the subtext for the images in the HTML and then put each
image with accompanying subtext into their own classes. Add a
class to all images to later create a flexbox around them. Add
different text and color properties to <li> and .header-logo to
better match the given example.