Add flex properites to hero and alter button
Add .hero-content to HTML to separate text and image into
different containers. Use flex properties to place text and
image side-by-side and to arrange them in relation to the rest
of the page and each other. Change <body> margin to make content
go all the way to the edge of the page.

Add rounded corners and padding to button.